+++ R/analysis_pred.R
@@ -109,7 +109,7 @@
 #' }
 #' 
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' data(microcredit)
 #' covars <- c("age", "gender", "education", "hhinc_yrly_base",
 #'             "css_creditscorefinal")